GMT_DOS-CLIENTS_ARROW SPLIT CRATE IN MODULES
Replace `get` module with `iter`
