+++ Cargo.toml
@@ -38,7 +38,7 @@ gmt_dos-clients_mount = { version = "3.0", default-features = false, path = "cli
 gmt_dos-clients_m1-ctrl = { version = "3.1", path = "clients/m1-ctrl/" }
 gmt_dos-clients_m2-ctrl = { version = "1.2", path = "clients/m2-ctrl/" }
 gmt_dos-clients_servos = { version = "1.0", default-features = false, path = "clients/servos/" }
-gmt_dos-clients_arrow = { version = "2.0", path = "clients/arrow/" }
+gmt_dos-clients_arrow = { version = "2.1", path = "clients/arrow/" }
 gmt_dos-clients_domeseeing = { version = "2.1", path = "clients/domeseeing/" }
 gmt_dos-clients_windloads = { version = "2.0.1", path = "clients/windloads/" }
 gmt_dos-clients_lom = { version = "1.1", path = "clients/lom/" }
