+++ clients/arrow/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "gmt_dos-clients_arrow"
-version = "2.0.1"
+version = "2.1.0"
 authors.workspace = true
 edition.workspace = true
 license.workspace = true
